7.13.26

+++ Hardware/DS157心率传感器/程序和流程图/DS157/Driver/tim_drive.c
@@ -21,7 +21,7 @@ void tim_init(TIM_Handle_t* htim)
     htim->ticks = 0;
 }
 
-float tim_get(TIM_Handle_t* htim)
+u16 tim_get(TIM_Handle_t* htim)
 {
     if (!htim)
 		return;
